Light celan up before SliverTree
