+++ packages/flutter/lib/src/rendering/sliver_fixed_extent_list.dart
@@ -466,3 +466,28 @@ class RenderSliverFixedExtentList extends RenderSliverFixedExtentBoxAdaptor {
     markNeedsLayout();
   }
 }
+
+/// A sliver that places multiple box children with the corresponding main axis extent in
+/// a linear array.
+class RenderSliverVariedExtentList extends RenderSliverFixedExtentBoxAdaptor {
+  /// Creates a sliver that contains multiple box children that have a explicit
+  /// extent in the main axis.
+  RenderSliverVariedExtentList({
+    required super.childManager,
+    required ItemExtentBuilder itemExtentBuilder,
+  }) : _itemExtentBuilder = itemExtentBuilder;
+
+  @override
+  ItemExtentBuilder get itemExtentBuilder => _itemExtentBuilder;
+  ItemExtentBuilder _itemExtentBuilder;
+  set itemExtentBuilder(ItemExtentBuilder value) {
+    if (_itemExtentBuilder == value) {
+      return;
+    }
+    _itemExtentBuilder = value;
+    markNeedsLayout();
+  }
+
+  @override
+  double? get itemExtent => null;
+}
