+++ packages/flutter/lib/src/widgets/scroll_view.dart
@@ -24,7 +24,6 @@ import 'scrollable.dart';
 import 'scrollable_helpers.dart';
 import 'sliver.dart';
 import 'sliver_prototype_extent_list.dart';
-import 'sliver_varied_extent_list.dart';
 import 'viewport.dart';
 
 // Examples can assume:
